Add data-property value to span element in card
To allow identification of the book property each span element
represents in the card, a data attribute is added to each element
with the property value.
The property text displayed on the webpage is returned by a new
function so that the bookProperties array may contain the actual
property names (in the code) as strings.