Display book details when creating a card
The createCard function now takes a book object as a parameter,
and for each book property displays the associated value to the
webpage.
Function also created to loop through an array of book objects and
display them as cards - this function will be used in future to
display all previously stored books in the library array.